moved classes and functions from script.js to their own separate files
